Updated Python
Renamed Python library, also implemented a temperature monitoring for the pi, able to show real time data of voltage for 2 minutes as well as displaying errors and feedback relating to the extraction and storage of firmware/voltage info.
