+++ python-test/testing_functions/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "nodemailer": "^6.9.14"
+  }
+}
